requirements.txt
dspy-ai
fastapi
uvicorn
pydantic
python-dotenv
